Jar Creation
- Made Modifications to have a single FXMLLoader instead of several.
This was done, as multiple loaders caused a bug when attempting to run
the Jar File.

- Also Updated the Video/Hyperlink File Paths. Now when you go to import
a Video, Save a Hyperlink File, or Open a Hyperlink file, the default
Location is the Computer System's Desktop. Will allow for commonality.

NOTE: REMOVING all video files and hyperlink files from repository.
Created appropriate folders on Computer's Desktop.